Removed/Refactored duplicated classes related to serialization in mlr and nmf
